+++ common/src/main/java/dev/terminalmc/flashside/gui/screen/ClothScreenProvider.java
@@ -18,6 +18,7 @@ package dev.terminalmc.flashside.gui.screen;
 
 import dev.terminalmc.flashside.config.Config;
 import me.shedaniel.clothconfig2.api.*;
+import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.screens.Screen;
 
 import static dev.terminalmc.flashside.util.Localization.localized;
@@ -38,44 +39,70 @@ public class ClothScreenProvider {
                 .setTitle(localized("name"))
                 .setSavingRunnable(Config::save);
         ConfigEntryBuilder eb = builder.entryBuilder();
-        
+
         ConfigCategory common = builder.getOrCreateCategory(localized("option", "common"));
 
-        common.addEntry(eb.startBooleanToggle(
-                        localized("option", "common.leftSide"), options.leftSide)
+        common.addEntry(eb.startBooleanToggle(localized("option", "common.leftSide"),
+                        options.leftSide)
                 .setTooltip(localized("option", "common.leftSide.tooltip"))
                 .setDefaultValue(Config.Options.leftSideDefault)
                 .setSaveConsumer(val -> options.leftSide = val)
                 .build());
-        
-        common.addEntry(eb.startBooleanToggle(
-                        localized("option", "common.modmenuIconTop"), options.modmenuIconTop)
+
+        common.addEntry(eb.startBooleanToggle(localized("option", "common.modmenuIconTop"),
+                        options.modmenuIconTop)
                 .setTooltip(localized("option", "common.modmenuIconTop.tooltip"))
                 .setDefaultValue(Config.Options.modmenuIconTopDefault)
                 .setSaveConsumer(val -> options.modmenuIconTop = val)
                 .build());
-        
+
         ConfigCategory pause = builder.getOrCreateCategory(localized("option", "pause"));
-        
-        pause.addEntry(eb.startIntSlider(
-                localized("option", "pause.startRow"), options.startRow, 0, 4)
+
+        pause.addEntry(eb.startIntSlider(localized("option", "pause.startRow"),
+                        options.startRow, 0, 4)
                 .setTooltip(localized("option", "pause.startRow.tooltip"))
                 .setDefaultValue(Config.Options.startRowDefault)
                 .setSaveConsumer(val -> options.startRow = val)
                 .setTextGetter(val -> localized("option", "pause.startRow.value", val)) // op
                 .build());
 
-        ConfigCategory title = builder.getOrCreateCategory(localized("option", "title"));
+        pause.addEntry(eb.startBooleanToggle(localized("option", "pause.commandEnabled"),
+                        options.commandEnabled)
+                .setTooltip(localized("option", "pause.commandEnabled.tooltip"))
+                .setDefaultValue(Config.Options.commandEnabledDefault)
+                .setSaveConsumer(val -> options.commandEnabled = val)
+                .build());
         
-        title.addEntry(eb.startBooleanToggle(
-                        localized("option", "title.editTitleScreen"), options.editTitleScreen)
-                .setTooltip(localized("option", "title.editTitleScreen.tooltip"))
+        pause.addEntry(eb.startStrField(localized("option", "pause.commandString"),
+                        options.commandString)
+                .setTooltip(localized("option", "pause.commandString.tooltip"))
+                .setDefaultValue(Config.Options.commandStringDefault)
+                .setSaveConsumer(val -> options.commandString = val)
+                .build());
+
+        for (int i = 0; i < options.actions.length; i++) {
+            int finalI = i;
+            pause.addEntry(eb.startEnumSelector(localized("option", "pause.action." + i),
+                            Config.Action.class, options.actions[i])
+                    .setDefaultValue(Config.Options.actionsDefault.get()[i])
+                    .setSaveConsumer(val -> options.actions[finalI] = val)
+                    .setEnumNameProvider((action) -> localized("option", "action." + action))
+                    .build());
+        }
+
+        ConfigCategory title = builder.getOrCreateCategory(localized("option", "title"));
+
+        title.addEntry(eb.startBooleanToggle(localized("option", "title.editTitleScreen"),
+                        options.editTitleScreen)
+                .setTooltip(localized("option", "title.editTitleScreen.tooltip").append("\n")
+                        .append(localized("option", "title.editTitleScreen.tooltip.warning")
+                                .withStyle(ChatFormatting.RED)))
                 .setDefaultValue(Config.Options.editTitleScreenDefault)
                 .setSaveConsumer(val -> options.editTitleScreen = val)
                 .build());
 
-        title.addEntry(eb.startIntSlider(
-                        localized("option", "title.startRowTitleScreen"), options.startRowTitleScreen, 0, 4)
+        title.addEntry(eb.startIntSlider(localized("option", "title.startRowTitleScreen"),
+                        options.startRowTitleScreen, 0, 4)
                 .setTooltip(localized("option", "title.startRowTitleScreen.tooltip"))
                 .setDefaultValue(Config.Options.startRowTitleScreenDefault)
                 .setSaveConsumer(val -> options.startRowTitleScreen = val)